Treat all DB dates/times as UTC
In reality of course they are *not* UTC—they're the local time of
the database server—but if we treat them as UTC then we avoid any
conversion headaches and can garuantee that the time displayed is
the time the event occurred _local to where it occurred_.

For our particular setup this means that the times displayed will
always reflect the BST adjusted time (though the database doesn't
include the offset to indicate such.) If we take the times as given
and don't perform any offsetting then they'll be displayed as our
users expect them to be.

The main change here is actually in the client code: using the
`getUTC...` variants of the date/time part extraction methods.
This avoids converting the time into client local time which (and
a potential “doubling up” of the BST offset.)