Remove Hysterix dependency
It had started to cause issues around query timeouts and given the
complexity of configuring it to suit our exact needs versus the limited
benefit it was providing removing it was the more sensible choice.

We don't really have to worry overmuch about database connectivity
issues and if there ever are any we can rely on the entire process being
restarted rather than any in-process circuit breaking.